+++ owncloudApp/src/main/java/com/owncloud/android/providers/FileContentProvider.kt
@@ -54,12 +54,9 @@ import com.owncloud.android.data.ProviderMeta
 import com.owncloud.android.data.capabilities.datasources.implementation.OCLocalCapabilitiesDataSource
 import com.owncloud.android.data.capabilities.datasources.mapper.OCCapabilityMapper
 import com.owncloud.android.data.capabilities.db.OCCapabilityEntity
-import com.owncloud.android.data.files.datasources.implementation.OCLocalFileDataSource
 import com.owncloud.android.data.files.db.FileDao
 import com.owncloud.android.data.files.db.OCFileEntity
-import com.owncloud.android.data.sharing.shares.datasources.implementation.OCLocalShareDataSource
 import com.owncloud.android.data.sharing.shares.db.OCShareEntity
-import com.owncloud.android.datamodel.OCFile
 import com.owncloud.android.datamodel.UploadsStorageManager
 import com.owncloud.android.db.ProviderMeta.ProviderTableMeta
 import com.owncloud.android.lib.common.accounts.AccountUtils
@@ -70,6 +67,7 @@ import java.io.File
 import java.io.FileNotFoundException
 import java.util.ArrayList
 import java.util.HashMap
+import com.owncloud.android.datamodel.OCFile as OCFileLegacy
 
 /**
  * The ContentProvider for the ownCloud App.
@@ -532,7 +530,7 @@ class FileContentProvider(val executors: Executors = Executors()) : ContentProvi
         return results
     }
 
-    private inner class DataBaseHelper internal constructor(context: Context?) :
+    private inner class DataBaseHelper(context: Context?) :
         SQLiteOpenHelper(
             context,
             ProviderMeta.DB_NAME,
@@ -1077,7 +1075,7 @@ class FileContentProvider(val executors: Executors = Executors()) : ContentProvi
                     // Insert share list to the new shares table in new database
                     val ocFileDao: FileDao by inject()
                     executors.diskIO().execute {
-                        for (file in files){
+                        for (file in files) {
                             ocFileDao.mergeRemoteAndLocalFile(file)
                         }
                     }
@@ -1343,7 +1341,7 @@ class FileContentProvider(val executors: Executors = Executors()) : ContentProvi
                     val oldPath = it.getString(
                         it.getColumnIndex(ProviderTableMeta.FILE_STORAGE_PATH)
                     )
-                    val file = OCFile(
+                    val file = OCFileLegacy(
                         it.getString(it.getColumnIndex(ProviderTableMeta.FILE_PATH))
                     )
                     val newPath = FileStorageUtils.getDefaultSavePathFor(newAccountName, file)
